feat: put hooks into separate package
