+++ packages/nextjs/package.json
@@ -31,7 +31,7 @@
     "@types/node": "^17.0.32",
     "@types/object-hash": "^1.3.4",
     "@types/react": "^18.0.15",
-    "@wundergraph/sdk": "workspace:^0.108.0",
+    "@wundergraph/sdk": "workspace:*",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "typescript": "^4.8.2"
