+++ packages/swr-hooks/README.md
@@ -1,0 +1,59 @@
+# WunderGraph SWR Integration
+
+![wunderctl](https://img.shields.io/npm/v/@wundergraph/swr-hooks.svg)
+
+This package provides a type-safe integration of [SWR](https://swr.vercel.app/) with WunderGraph.
+SWR is a React Hooks library for data fetching. With just one hook, you can significantly simplify the data fetching logic in your project. And it also covered in all aspects of speed, correctness, and stability to help you build better experiences.
+
+> **Warning**: Only works with WunderGraph.
+
+## Getting Started
+
+```shell
+npm install @wundergraph/swr-hooks
+```
+
+Before you can use the hooks, you need to modify your code generation to include the base typescript client.
+
+```typescript
+// wundergraph.config.ts
+configureWunderGraphApplication({
+  // ... omitted for brevity
+  codeGenerators: [
+    {
+      templates: [templates.typescript.client, ...templates.typescript.models],
+      // the location where you want to generate the client
+      path: '../src/components/generated',
+    },
+  ],
+});
+```
+
+Second, run `wunderctl generate` to generate the code.
+
+Now you can use the hooks.
+
+```ts
+import { newSWRHooks } from '@wundergraph/swr-hooks';
+import { createClient, Mutations, Queries, Subscriptions } from './components/generated/client';
+const { useQuery, useMutation } = newSWRHooks<Queries, Mutations, Subscriptions>(createClient());
+
+const { error, data } = useQuery({
+  operationName: 'Weather',
+  input: { forCity: city },
+  liveQuery: true,
+});
+
+const { mutate } = useMutation({
+  operationName: 'SetName',
+});
+mutate({
+  input: { name },
+});
+```
+
+## Options
+
+You can use all available options from [SWR](https://swr.vercel.app/docs/options) with the hooks.
+Due to the fact that we use the operationName + variables as **key**, you can't use the `key` option as usual.
+In order to use [conditional-fetching](https://swr.vercel.app/docs/conditional-fetching) you can use the `enabled` option.
