+++ packages/swr-hooks/src/index.ts
@@ -1,0 +1,3 @@
+export { newSWRHooks } from './hooks';
+
+export type { UseMutationOptions, UseSubscriptionOptions, MutateOptions, UseQueryOptions } from './hooks';
