+++ packages/vite-testapp/package.json
@@ -15,9 +15,9 @@
   },
   "dependencies": {
     "@wundergraph/sdk": "workspace:*",
+    "@wundergraph/swr-hooks": "workspace:*",
     "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "swr": "^1.3.0"
+    "react-dom": "^18.1.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.15",
