+++ packages/vite-testapp/src/App.tsx
@@ -1,7 +1,10 @@
 import reactLogo from './assets/react.svg';
 import './App.css';
-import { useQuery, useMutation } from './hooks/swr';
+import { newSWRHooks } from '@wundergraph/swr-hooks';
 import { useState } from 'react';
+
+import { createClient, Mutations, Queries, Subscriptions } from './components/generated/client';
+const { useQuery, useMutation } = newSWRHooks<Queries, Mutations, Subscriptions>(createClient());
 
 const LiveWeather: React.FC<{ city: string }> = ({ city }) => {
 	const liveWeather = useQuery({
