+++ packages/vite-testapp/vite.config.ts
@@ -5,11 +5,11 @@ import react from '@vitejs/plugin-react';
 export default defineConfig({
 	plugins: [react()],
 	optimizeDeps: {
-		include: ['@wundergraph/sdk', '@wundergraph/sdk/internal', '@wundergraph/sdk/client'],
+		include: ['@wundergraph/sdk', '@wundergraph/swr-hooks', '@wundergraph/sdk/client'],
 	},
 	build: {
 		commonjsOptions: {
-			include: [/sdk\/dist\//, /node_modules/, /@wundergraph/],
+			include: [/sdk\/dist\//, /node_modules/, /wundergraph/],
 		},
 	},
 });
